Add the mapstruct to pom.xml with errors in mapper package

+++ src/main/java/com/fawry/librarysystem/service/AuthorService.java
@@ -1,49 +1,63 @@
 package com.fawry.librarysystem.service;
 
-import com.fawry.librarysystem.entity.Author;
-import com.fawry.librarysystem.entity.Book;
+import com.fawry.librarysystem.mapper.AuthorMapper;
+import com.fawry.librarysystem.mapper.BookMapper;
+import com.fawry.librarysystem.model.dto.AuthorDTO;
+import com.fawry.librarysystem.model.dto.BookDTO;
 import com.fawry.librarysystem.repository.AuthorRepo;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class AuthorService {
-    AuthorRepo authorRepo;
+    private final AuthorRepo authorRepo;
+    private final AuthorMapper authorMapper;
+    private final BookMapper bookMapper;
 
-    public AuthorService(AuthorRepo authorRepo) {
-        this.authorRepo = authorRepo;
+    public void addAuthor(AuthorDTO author) {
+        if(author.getId() != null && authorRepo.findById(author.getId()).isPresent()) {
+            throw new RuntimeException("Author ID must be null");
+        }
+
+        authorRepo.save(authorMapper.toEntity(author));
     }
 
-    public void addAuthor(Author author) {
-        authorRepo.save(author);
-    }
+    public void updateAuthor(AuthorDTO author) {
+        if(author.getId() == null || !authorRepo.findById(author.getId()).isPresent()) {
+            throw new RuntimeException("Can't find the author");
+        }
 
-    public Author updateAuthor(Author author) {
-        return authorRepo.save(author);
+        authorMapper.toDto(authorRepo.save(authorMapper.toEntity(author)));
     }
 
     public void deleteAuthor(Long id) {
+        if(!authorRepo.findById(id).isPresent()) {
+            throw new RuntimeException("Can't find the author");
+        }
+
         authorRepo.deleteById(id);
     }
 
-    public void deleteAuthor(String name) {
-        authorRepo.deleteByName(name);
+    public AuthorDTO findAuthorById(Long id) {
+        if(!authorRepo.findById(id).isPresent()) {
+            throw new RuntimeException("Can't find the author");
+        }
+
+        return authorMapper.toDto(authorRepo.findById(id).get());
     }
 
-    public Author findAuthorById(Long id) {
-        return authorRepo.findById(id).orElse(null);
+    public List<AuthorDTO> findAllAuthors() {
+        return authorMapper.toDto(authorRepo.findAll());
     }
 
-    public List<Author> findAllAuthors() {
-        return authorRepo.findAll();
-    }
+    public List<BookDTO> findBooksByAuthorId(Long id) {
+        if(!authorRepo.findById(id).isPresent()) {
+            throw new RuntimeException("Can't find the author");
+        }
 
-    public Author findAuthorByName(String name) {
-        return authorRepo.findByName(name);
-    }
-
-    public List<Book> findAuthorsByBookId(Long id) {
-        return authorRepo.findAuthorsByBookId(id);
+        return bookMapper.toDto(authorRepo.findBooksByAuthorId(id));
     }
 }
